+++ src/bp/ml/svm-js/grid-search/typings.ts
@@ -1,0 +1,8 @@
+import { SvmParameters, Report } from '../typings'
+
+export type GridSearchResult = { params: SvmParameters; report: Report }
+
+export type GridSearchProgress = {
+  done: number
+  total: number
+}
