+++ src/bp/ml/svm-js/kernel-types.ts
