+++ src/bp/ml/svm-js/normalize-input.ts
