+++ src/bp/ml/svm.ts
@@ -2,10 +2,10 @@ import * as sdk from 'botpress/sdk'
 import _ from 'lodash'
 
 import { Data, SvmModel, SvmParameters as Parameters } from './svm-js/typings'
-import { SVM } from './svm-js/core/svm'
-import svmTypes from './svm-js/core/svm-types'
-import kernelTypes from './svm-js/core/kernel-types'
-import { getMinKFold } from './svm-js/util/split-dataset'
+import { SVM } from './svm-js/svm'
+import svmTypes from './svm-js/svm-types'
+import kernelTypes from './svm-js/kernel-types'
+import { getMinKFold } from './svm-js/grid-search/split-dataset'
 
 export const DefaultTrainArgs: Partial<sdk.MLToolkit.SVM.SVMOptions> = {
   c: [0.1, 1, 2, 5, 10, 20, 100],
